ci: fix docker build caching for operator image
The caching bug occurred because all build arguments were set at the top of the dockerfile.
But build arguments invalidate the cache as soon as a RUN step comes up since they are passed along to the cmd
in the RUN step as env vars and docker doesn't know what happens inside the cmd, so the cache needs to be invalidated.

By reordering the ARG steps to only set the build arguments directly before they are used, I achieved far better caching,
especially for go mod download and installation of the controller tools.

Signed-off-by: Moritz Wiesinger <moritz.wiesinger@dynatrace.com>