feat: wire the new cert-manager into lifecycle operator (#529)
BREAKING CHANGES: The dependency on cert-manager has been removed in favor of a custom implementation. With these changes, the operator will be waiting for a certificate to be ready before registering the controllers. The certificate is generated as a k8s secret in the lifecycle-toolkit namespace and then loaded into an empty dir volume. The Keptn certificate manager will make sure to renew it (every 6 hours) and will take care of its validity every time the controller manager deployment resource changes.