docs: change Development url
We thought we could have a nicer url with `dev.` but it seems
like this is not working (or at least we did not manage).

As this is not mission critical we change it to simply use the
branch name, which works, and in fact needs lesser configuration.

Additionally fixing other urls too
Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>