fix: htmltest error for newly created documents
HTMLTest will complain, when we create new documentation pages,
as they are not within the git repository. As we used to display
the edit and create issue links pointing to the main branch, our
build would fail.

With this change, we will deactivate the rendering of those edit
links, when invoked via make target. This means locally and with
GitHub Actions, we will not see them. But netlify etc will
render them properly.

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>